+++ src/sm/refinedmesh.h
@@ -58,20 +58,20 @@ public:
 
 private:
     typedef struct fine_edge_rec {
-        long *fine_id;
+        int *fine_id;
     } fine_edge_rec;
 
     typedef struct fine_quad_rec {
-        long *fine_id;
+        int *fine_id;
     } fine_quad_rec;
 
     typedef struct fine_hexa_rec {
-        long *fine_id;
+        int *fine_id;
     } fine_hexa_rec;
 
 
     typedef struct fe_node_rec {
-        long id;
+        int id;
     } fe_node_rec;
 
     typedef struct fe_edge_rec {
@@ -99,43 +99,43 @@ private:
     typedef struct mesh_edge_rec {
         fe_node_rec *node [ 2 ];
         fe_node_rec *mid_node;
-        long *fine_id [ 2 ];
+        int *fine_id [ 2 ];
     } mesh_edge_rec;
 
     typedef struct mesh_face_rec {
         fe_node_rec *node [ 3 ];
         fe_node_rec *mid_node;
-        long *fine_id [ 3 ];
+        int *fine_id [ 3 ];
     } mesh_face_rec;
 
     typedef struct mesh_quad_rec {
         fe_node_rec *node [ 4 ];
         fe_node_rec *mid_node;
-        long *fine_id [ 4 ];
+        int *fine_id [ 4 ];
     } mesh_quad_rec;
 
 
 
 #ifdef COMPLETE_DATA_STRUCTURE
     typedef struct tmp_face_rec {
-        long ngb_elem_id [ 3 ];
+        int ngb_elem_id [ 3 ];
         mesh_edge_rec *edge [ 3 ];
     } tmp_face_rec;
 
     typedef struct tmp_quad_rec {
-        long ngb_elem_id [ 4 ];
+        int ngb_elem_id [ 4 ];
         mesh_edge_rec *edge [ 4 ];
     } tmp_quad_rec;
 #endif
 
 
     typedef struct tmp_tetra_rec {
-        long ngb_elem_id [ 4 ];
+        int ngb_elem_id [ 4 ];
         mesh_face_rec *face [ 4 ];
     } tmp_tetra_rec;
 
     typedef struct tmp_hexa_rec {
-        long ngb_elem_id [ 6 ];
+        int ngb_elem_id [ 6 ];
         mesh_quad_rec *quad [ 6 ];
     } tmp_hexa_rec;
